finished with find-max-value method from binary tree class
